PCMSampleSource tested. Tracing AudioPlayer source_ interface bugs.

+++ experiments/audio-test/Audio/Inc/audio_sink.h
@@ -8,7 +8,7 @@ public:
     virtual bool writeSample(int16_t sample) = 0;
     virtual bool stop() = 0;
     virtual bool silence() = 0;
-    virtual bool prepare(uint32_t sampleRateHz) = 0;
+    virtual bool prepare(uint16_t sampleRateHz) = 0;
     virtual bool start() = 0;
     virtual bool isSampleNeeded() const = 0;
 };
